Refactor : Common 패키지 구성 변경 (ApiUtil 과 Common****Handler)
