+++ order-server/src/main/java/com/inbobwetrust/aop/ExceptionHandlerControllerAdvice.java
@@ -1,6 +1,6 @@
-package com.inbobwetrust.common;
+package com.inbobwetrust.aop;
 
-import static com.inbobwetrust.common.ApiUtil.errorResponse;
+import static com.inbobwetrust.aop.ApiResult.errorResponse;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
 @ControllerAdvice
-public class CommonExceptionHandler {
+public class ExceptionHandlerControllerAdvice {
     private final Logger log = LoggerFactory.getLogger(getClass());
 
     @ExceptionHandler(value = {IllegalArgumentException.class})
