--- order-server/src/main/java/com/inbobwetrust/common/ApiResult.java
@@ -1,11 +1,0 @@
-package com.inbobwetrust.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ApiResult<T> {
-    private final boolean success;
-    private final T body;
-}
