+++ order-server/src/main/java/com/inbobwetrust/config/swaggerdoc/RiderControllerSwaggerDoc.java
@@ -1,6 +1,6 @@
 package com.inbobwetrust.config.swaggerdoc;
 
-import com.inbobwetrust.common.ApiResult;
+import com.inbobwetrust.aop.ApiResult;
 import com.inbobwetrust.model.vo.Rider;
 
 import io.swagger.annotations.ApiOperation;
